[jk] Differentiate current_user local storage (#5294)
# Description
- Allow user with multiple Mage accounts to sign into different clusters
on the same browser at the same time.
- There was an issue with the local storage clearing user profile data
for a different account when the user signed out, so this PR makes the
current user local storage key unique to the Mage deployment's base
path. This doesn't affect users that use a single Mage deployment on
their browser.

# How Has This Been Tested?
- Used different accounts to sign into multiple Mage cluster deployments
with different base paths (but same domain). Signing out of one
deployment did not affect the other user profile in local storage.

# Checklist
- [X] The PR is tagged with proper labels (bug, enhancement, feature,
documentation)
- [X] I have performed a self-review of my own code
- [ ] I have added unit tests that prove my fix is effective or that my
feature works
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation

cc:
<!-- Optionally mention someone to let them know about this pull request
-->