[jk] Update condition for cookie secure property (#5327)
# Description
- The `secure` cookie property was causing login issues if the user was
running Mage not on localhost and not with `https`. This PR updates the
condition that was setting the `secure` property to `true`. Previously,
it checked if the `NODE_ENV` environment variable was `production`, but
the FE production build could still have a "production" `NODE_ENV`
environment variable set while still using the `http` protocol.

# How Has This Been Tested?
- Tested locally in Chrome, Safari, and Firefox.

# Checklist
- [X] The PR is tagged with proper labels (bug, enhancement, feature,
documentation)
- [X] I have performed a self-review of my own code
- [ ] I have added unit tests that prove my fix is effective or that my
feature works
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation

cc:
<!-- Optionally mention someone to let them know about this pull request
-->