[xy] Disable access after user is deleted. (#5867)
# Description
<!-- Please include a summary of the change and which issue is fixed.
Please also include relevant motivation and context.
List any dependencies that are required for this change.
-->
Fix CVE: https://nvd.nist.gov/vuln/detail/CVE-2024-45187

This pull request introduces a small but important change to the user
authentication logic in the terminal server. The update ensures that if
a user is deleted, their access is immediately revoked by setting the
`valid` flag to `False`.

* Authentication logic improvement:
*
[`mage_ai/server/terminal_server.py`](diffhunk://#diff-223660bbbc0aa2187ec2f4af856e298ba37373323b92d253f24d35d6cfb17067R129-R131):
Added a check to set `valid = False` if the user is deleted, preventing
unauthorized access to the terminal.

# How Has This Been Tested?
<!-- Please describe the tests that you ran to verify your changes.
Provide instructions so we can reproduce.
-->

- [x] Tested locally


# Checklist
- [x] The PR is tagged with proper labels (bug, enhancement, feature,
documentation)
- [x] I have performed a self-review of my own code
- [ ] I have added unit tests that prove my fix is effective or that my
feature works
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation

cc:
<!-- Optionally mention someone to let them know about this pull request
-->