Making projects to be ordered by their names alphabetically in the fr… (#5918)
…ee tier

# Description
This PR is dedicated to add a change which makes projects to be ordered
by their names alphabetically in the free tier.


# How Has This Been Tested?
I was testing it changing some parts of the code base. After I reached
what I wanted, I reduced the amount of changes to the necessary ones
only. The change affects 3 rows only, the effect of which is easy to
see. In case if the not recommended multi-project approach (free tier),
the project selector was ordered alphabetically as wanted.

# Checklist
- [ ] The PR is tagged with proper labels (bug, enhancement, feature,
documentation)
- [ ] I have performed a self-review of my own code
- [ ] I have added unit tests that prove my fix is effective or that my
feature works
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation

@wangxiaoyou1993

Co-authored-by: Andrejszki Alex <alex.andrejszki@theitsolutions.io>