[xy] Lock package versions (#5939)
# Description
<!-- Please include a summary of the change and which issue is fixed.
Please also include relevant motivation and context.
List any dependencies that are required for this change.
-->
 Lock package versions

This pull request updates the dependency specifications for `langchain`,
`langchain_community`, and `openai` in the `setup.py` file. The changes
ensure that these packages use exact versions rather than allowing any
version greater than the minimum specified, which helps maintain
compatibility and reproducibility.

Dependency version pinning:

* Changed `langchain` and `langchain_community` requirements from
`>=0.2.5` to `==0.2.5` in both the `ai` and main `install_requires`
extras to enforce the use of a specific version.
[[1]](diffhunk://#diff-60f61ab7a8d1910d86d9fda2261620314edcae5894d5aaa236b821c7256badd7L45-R47)
[[2]](diffhunk://#diff-60f61ab7a8d1910d86d9fda2261620314edcae5894d5aaa236b821c7256badd7L180-R185)
* Changed `openai` requirement from `>=1.0.0` to `==1.82.0` in both the
`ai` and main `install_requires` extras to enforce the use of a specific
version.
[[1]](diffhunk://#diff-60f61ab7a8d1910d86d9fda2261620314edcae5894d5aaa236b821c7256badd7L45-R47)
[[2]](diffhunk://#diff-60f61ab7a8d1910d86d9fda2261620314edcae5894d5aaa236b821c7256badd7L180-R185)

# How Has This Been Tested?
<!-- Please describe the tests that you ran to verify your changes.
Provide instructions so we can reproduce.
-->

- [x] Tested locally and in CI


# Checklist
- [x] The PR is tagged with proper labels (bug, enhancement, feature,
documentation)
- [x] I have performed a self-review of my own code
- [ ] I have added unit tests that prove my fix is effective or that my
feature works
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation

cc:
<!-- Optionally mention someone to let them know about this pull request
-->